backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
openai==1.3.5
